+++ src/Endpoints/PaymentChargebackEndpoint.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace Mollie\Api\Endpoints;
+
+use Mollie\Api\Resources\Chargeback;
+use Mollie\Api\Resources\ChargebackCollection;
+use Mollie\Api\Resources\Payment;
+
+class PaymentChargebackEndpoint extends EndpointAbstract
+{
+    protected $resourcePath = "payments_chargebacks";
+
+    /**
+     * Get the object that is used by this API endpoint. Every API endpoint uses one type of object.
+     *
+     * @return Chargeback
+     */
+    protected function getResourceObject()
+    {
+        return new Chargeback($this->api);
+    }
+
+    /**
+     * Get the collection object that is used by this API endpoint. Every API endpoint uses one type of collection object.
+     *
+     * @param int $count
+     * @param object[] $_links
+     *
+     * @return ChargebackCollection
+     */
+    protected function getResourceCollectionObject($count, $_links)
+    {
+        return new ChargebackCollection($this->api, $count, $_links);
+    }
+
+    /**
+     * @param Payment $payment
+     * @param string $chargebackId
+     * @param array $parameters
+     *
+     * @return Chargeback
+     */
+    public function getFor(Payment $payment, $chargebackId, array $parameters = [])
+    {
+        $this->parentId = $payment->id;
+
+        return parent::rest_read($chargebackId, $parameters);
+    }
+}
