+++ tests/Mollie/TestHelpers/AmountObjectTestHelpers.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Tests\Mollie\TestHelpers;
+
+trait AmountObjectTestHelpers
+{
+    protected function assertAmountObject($value, $currency, $amountObject)
+    {
+        return $this->assertEquals(
+            $this->createAmountObject($value, $currency),
+            $amountObject
+        );
+    }
+
+    protected function createAmountObject($value, $currency)
+    {
+        return (object) [
+            'value' => $value,
+            'currency' => $currency,
+        ];
+    }
+}
